Ensure that ERRORs are handled as well as SERIOUS-CONDITIONs
The *DEBUGGER-HOOK* binding added in c21df22 was being preempted by
the active ERROR handler introduced by the outermost HANDLER-CASE in
%RPC-SERVER-THREAD-WORKER. As a result, any ERRORs signaled during the
processing (as opposed to SERIOUS-CONDITIONs) wound up in the "this is
where errors go where we can't even reply to the client" handler,
which only logged a message but did not send any reply. This would
cause clients to hang indefinitely awaiting a response.

As a quick-fix: (1) switch the outer HANDLER-CASE to a TAGBODY form
that establishes a :START tag at the top of the loop, (2) push the
HANDLER-CASE scope down to only cover the call to %PULL-RAW-REQUEST,
and (3) (GO :START) in the error handler after logging the error.

Since there is no longer any active, applicable handler for ERRORs
that occur within the scope of the *DEBUGGER-HOOK* binding,
INVOKE-DEBUGGER is called on our custom ERROR-PROCESSOR, which handles
logging and creating the |RPCError| response.

See also the discussion in https://github.com/rigetti/rpcq/pull/94