Coerce *args to a VECTOR in PREPARE-RPC-CALL-ARGS
Coerce positional args to a VECTOR in PREPARE-RPC-CALL-ARGS to ensure
they are encoded as an array. This makes no difference when positional
args exist since both non-empty LISTs and ARRAYs get encoded to
messagepack array formats; if args is NIL, however, CL-MESSAGEPACK
encodes NIL to a messagepack NIL format 0xc0, which gets decoded as
None on the Python side. This results in an Exception being raised in
RPCSpec.run_handler when attempting to decode the args since it
expects *args to be an Iterable, not None.

The RPCQ lisp server was unaffected by this since messagepack NIL is
decode to lisp NIL, and the lisp server only calls generic SEQUENCE
processing functions (CONCATENATE) on the deserialized value, and
hence doesn't distinguish between an empty LIST or ARRAY.