Pass client timeout to server, optionally override server's timeout (#126)
* Pass client timeout to server, optionally override server's timeout

If the server's timeout is shorter than the client's, keep it. If the
client's is shorter than the server's, use that timeout for the
server. This prevents the server running jobs for which the client
will not receive the results because it has already timed-out.

Update docstrs and leave a note for future hackers

Para el Appleby

fix

Add test

* Use client timeout in RPC-CALL

* Improve test case

* Update src-tests/test-rpc.lisp

Co-Authored-By: appleby <mappleby@rigetti.com>

Co-authored-by: appleby <mappleby@rigetti.com>